Refactored to a RendererAnimatedSprite system.
